requirements.txt
# Core dependencies
agentica>=0.2.0
fastapi>=0.109.0
uvicorn>=0.27.0
websockets>=12.0
pydantic>=2.0.0
python-dotenv>=1.0.0
loguru>=0.7.0

# Channels
lark-oapi>=1.0.0        # Feishu
apscheduler>=3.10.0     # Scheduler
pyyaml>=6.0             # YAML config

# Testing
pytest>=7.0
pytest-asyncio>=0.23.0
httpx>=0.27.0           # Required by FastAPI TestClient

# Optional: Telegram
# python-telegram-bot>=20.0

# Optional: Discord
# discord.py>=2.0
